Fix mistake to send email at several recipients
When I use this plugin to send Notifications at several
recipients, only the first recipient receives the email.

The emails of recipients must not start or end with space(s)
otherwise the SMTP instruction "RCPT TO:< my@email.com>"
doesn't work.

The list of recipients transmitted to Message class can be
composed of email separated by ", " and not only ",". When
the list of recipients is splitted thanks to delimiter ","
the email can start by space and it must be removed.

Signed-off-by: Sébastien Mennetrier <smennetrier@voxtok.com>